Add option to choose which files to download

+++ khinsider/khscraper.py
@@ -1,6 +1,6 @@
 import requests, bs4, time, sys, json, os
 
-version = "1.9"
+version = "2.0"
 htmlFile = 'webpage.txt'
 cacheFile = 'musicData.json'
 linkPrefix = 'https://downloads.khinsider.com'
@@ -8,10 +8,14 @@ linkPrefix = 'https://downloads.khinsider.com'
 print("-- KHINSIDER SCRAPER --")
 print(f"V. {version}\n")
 
+def clearScreen():
+    """ Clears the terminal screen using the OS specific method.
+    """
+    os.system('clear' if sys.platform == 'linux' else 'cls')
+
 def textPrompt():
     """ Read user input and return it.
     """
-
     response = sys.stdin.readline()
     response = response.rstrip()
 
@@ -20,8 +24,7 @@ def textPrompt():
 def confirmationPrompt(promptText):
     """ Provide a yes/no prompt to the user.
     """
-
-    print(f"\t:: {promptText} (y/n)")
+    print(f":: {promptText} (y/n)")
     response = sys.stdin.readline()
     response = response.rstrip().lower()
 
@@ -33,7 +36,6 @@ def confirmationPrompt(promptText):
 def statusMessage(message, status=None, prefix="  ", suffix=""):
     """ Show a status message to the user, the current action is represented in brackets.
     """
-
     if status is not None:
         print(f"{prefix}[{status}] {message}{suffix}")
     else:
@@ -42,7 +44,6 @@ def statusMessage(message, status=None, prefix="  ", suffix=""):
 def cleanDir():
     """ Clean the directory of all temporary files.
     """
-
     if os.path.isfile(htmlFile):
         os.remove(htmlFile)
     if os.path.isfile(cacheFile):
@@ -53,8 +54,7 @@ def cleanDir():
 def downloadWebpage(link):
     """ Download and save a webpage as html.
     """
-
-    statusMessage("Downloading webpage...", status="download:webpage", prefix="\n")
+    statusMessage("Downloading webpage...", status="download:webpage", prefix="\n  ")
     response = requests.get(link)
     type(response)
     response.raise_for_status()
@@ -66,7 +66,6 @@ def downloadWebpage(link):
 def parseHtml():
     """ Read a saved html file and return a parsed object.
     """
-
     with open(htmlFile) as rawHtml:
         parsedHtml = bs4.BeautifulSoup(rawHtml.read(), features="html.parser")
 
@@ -75,7 +74,6 @@ def parseHtml():
 def scrapeHrefs(parsedHtml):
     """ Scrape a list of all tracks from the site.
     """
-
     selection = parsedHtml.select('.playlistDownloadSong a')
     type(selection)
 
@@ -97,7 +95,6 @@ def scrapeHrefs(parsedHtml):
 def scrapeSongLinks(trackList):
     """ Scrape a list of track links from the site.
     """
-
     musicList = []
     trackCount = len(trackList)
     count = 1
@@ -117,6 +114,7 @@ def scrapeSongLinks(trackList):
         parsedHtml = bs4.BeautifulSoup(res.text, features="html.parser")
         selection = parsedHtml.find('span', class_="songDownloadLink")
         music = {
+            'picked': True,
             'link': selection.parent['href'],
             'title': track['title'],
         }
@@ -131,36 +129,103 @@ def scrapeSongLinks(trackList):
 
     return musicList
 
+def listTracks(musicList):
+    """ Print tracks from a list
+    """
+    clearScreen()
+    for i, track in enumerate(musicList):
+        selection = 'x' if track['picked'] else ' '
+        print(f"[{selection}] {i+1} : {track['title']}")
+
+def chooser(musicList):
+    """ Exclude/Include tracks for download
+    """
+    print("Enter a comma separated list of numbers to exclude/include. Eg. (1,5,8,17)")
+    print("Items included will be excluded, items excluded will be included.")
+    choices = textPrompt().split(',')
+    clearScreen()
+
+    for item in choices:
+        try:
+            itemNumber = int(item)
+            isPicked = musicList[itemNumber - 1]['picked']
+            trackName = musicList[itemNumber - 1]['title']
+
+            selection = 'x' if not isPicked else ' '
+            musicList[itemNumber - 1]['picked'] = not isPicked
+
+            print(f"[{selection}] : {trackName}")
+        except IndexError as indexError:
+            print(f"Err: {indexError}")
+        except ValueError as valueError:
+            print(f"Err: {valueError}")
+
+    return musicList
+
+def trackPicker(musicList):
+    """ Displays a dialog allowing the user to choose what to download.
+    """
+    clearScreen()
+    userIsNotDone = True
+
+    while userIsNotDone:
+        print("\n:: Music Picker ::")
+        print("1) List tracks")
+        print("2) Exclude/Include tracks for download")
+        print("3) Done!")
+
+        print("What to do?")
+        response = textPrompt()
+        if response == '1':
+            listTracks(musicList)
+        elif response == '2':
+            musicList = chooser(musicList)
+        elif response == '3':
+            if confirmationPrompt("Are you sure?"):
+                userIsNotDone = False
+            clearScreen()
+        else:
+            print("Invalid option, try again.")
+
+    return musicList
+
 def downloadTracks(musicList):
     """ Download a list of links
     """
-
     trackCount = len(musicList)
+
     count = 1
     sleepTimer = 10
-    statusMessage("Starting...", status="download", prefix="\n", suffix="\n")
+    statusMessage("Starting...", status="download", prefix="\n  ", suffix="\n")
 
     for music in musicList:
-        statusMessage(f"({count} OF {trackCount}) | '{music['title']}'", prefix="")
+        if music['picked']:
+            statusMessage(f"({count} OF {trackCount}) | '{music['title']}'", prefix="")
+            download = requests.get(music['link'])
+            type(download)
+            download.raise_for_status()
 
-        download = requests.get(music['link'])
-        type(download)
-        download.raise_for_status()
+            with open(f"./downloadedTracks/{music['title']}.mp3", 'wb') as musicFile:
+                musicFile.write(download.content)
+            if count != trackCount:
+                time.sleep(sleepTimer)
+        else:
+            statusMessage(f"({count} OF {trackCount}) | '{music['title']}'", prefix="", suffix="[Not Downloaded]")
 
-        with open(f"./downloadedTracks/{music['title']}.mp3", 'wb') as musicFile:
-            musicFile.write(download.content)
-
-        if count != trackCount:
-            time.sleep(sleepTimer)
         count += 1
 
-    statusMessage("Complete!", status="download")
+    statusMessage("Complete!", status="download", prefix="\n  ")
     cleanDir()
 
 def main():
     if os.path.isfile(cacheFile) and confirmationPrompt("Found a cache file, download it?"):
+        # Load music list from json cache
         with open(cacheFile, 'r') as cache:
             musicList = json.load(cache)
+
+        if confirmationPrompt("Show track chooser?"):
+            musicList = trackPicker(musicList)
+
         downloadTracks(musicList)
     else:
         print("Paste link to download from khinsider: ")
@@ -169,6 +234,9 @@ def main():
         html = parseHtml()
         trackList = scrapeHrefs(html)
         musicList = scrapeSongLinks(trackList)
+
+        if confirmationPrompt("Show track chooser?"):
+            musicList = trackPicker(musicList)
 
         if confirmationPrompt("Download music now?"):
             downloadTracks(musicList)
